Fix type conversion errors changing int to size_t

+++ cpu/main.cpp
@@ -27,7 +27,7 @@ int main(int argc, char *argv[])
 
     std::string text = readStrFromFile(textFile);
     std::string keyword = readStrFromFile(keywordFile);
-    std::vector<int> results;
+    std::vector<size_t> results;
 
     const int iterNum = 20;
